// src/tokens.d.ts
type BackgroundKey = "critical" | "critical-active" | "critical-hover" | "critical-secondary" | "critical-secondary-weak" | "primary" | "primary-active" | "primary-disabled" | "primary-hover" | "secondary" | "secondary-active" | "secondary-disabled" | "secondary-hover" | "secondary-weak" | "success" | "success-active" | "success-hover" | "success-secondary" | "success-secondary-weak" | "surface-1" | "surface-1-hover" | "surface-1-inverse" | "surface-1-inverse-hover" | "surface-2" | "surface-2-hover" | "surface-3" | "surface-3-hover" | "warning" | "warning-secondary" | "warning-secondary-weak";